fixup! Do not overwrite existing config file.

+++ Makefile
@@ -25,8 +25,7 @@ install: uninstall docker-multipass-config docker-multipass \
 	ln -s $(LIB_DIR)/background-foreground $(LIB_DIR)/foreground
 	install $(CURDIR)/lib/docker-multipass/stop $(LIB_DIR)/stop
 	install $(CURDIR)/lib/docker-multipass/start $(LIB_DIR)/start
-	[ -f "$(HOME)/.docker-multipass-conf" ] || \
-		install $(CURDIR)/docker-multipass-config $(HOME)/.docker-multipass-conf
+	[ -f "$(HOME)/.docker-multipass-conf" ] || install $(CURDIR)/docker-multipass-config $(HOME)/.docker-multipass-config
 
 uninstall:
 	rm -f $(BINARY_DIR)/docker-multipass
